refactor: 패키지 변경, 방 도메인 코드 정리
