+++ src/main/java/prography/pingpong_game/common/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.config;
+package prography.pingpong_game.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
