+++ src/main/java/prography/pingpong_game/common/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.config;
+package prography.pingpong_game.common.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import org.springframework.context.annotation.Configuration;
