+++ src/main/java/prography/pingpong_game/common/controller/HealthController.java
@@ -1,10 +1,10 @@
-package prography.pingpong_game;
+package prography.pingpong_game.common.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import prography.pingpong_game.common.ApiResponse;
+import prography.pingpong_game.common.dto.ApiResponse;
 
 @RestController
 @RequestMapping("/health")
