+++ src/main/java/prography/pingpong_game/common/dto/ApiResponse.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.common;
+package prography.pingpong_game.common.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.v3.oas.annotations.media.Schema;
