+++ src/main/java/prography/pingpong_game/common/exception/GlobalExceptionHandler.java
@@ -4,7 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import prography.pingpong_game.common.ApiResponse;
+import prography.pingpong_game.common.dto.ApiResponse;
 
 @RestControllerAdvice
 @Slf4j
