+++ src/main/java/prography/pingpong_game/initialization/controller/InitializationController.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import prography.pingpong_game.common.ApiResponse;
+import prography.pingpong_game.common.dto.ApiResponse;
 import prography.pingpong_game.initialization.dto.request.InitializationRequest;
 import prography.pingpong_game.initialization.service.InitializationService;
