+++ src/main/java/prography/pingpong_game/room/event/GameEventListener.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.common.event;
+package prography.pingpong_game.room.event;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
