+++ src/main/java/prography/pingpong_game/room/event/GameStartEvent.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.common.event;
+package prography.pingpong_game.room.event;
 
 
 public record GameStartEvent(Long roomId) {
